requirements.txt
requests>=2.31.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sse-starlette>=1.8.2
python-multipart>=0.0.9
python-dotenv>=1.0.0
